Extract shared binary compatibility test helpers and refactor scalar integration tests to reuse them
